fix: quit confirm also honours the keymap, and up/down not just left/right
The quit-confirmation popup (src/tui/logic/input/quit.rs) has its own
hardcoded Left/Right/Up/Down toggle, separate from handle_confirm_input -
same bug, different popup, missed in the first pass.

Also broadens the keymap check itself: it only looked at SubTabLeft/
SubTabRight, but arrow keys toggle on all four directions, so a keymap-bound
Up/Down (default k/j) should too, not just Left/Right (h/l).

Pulled the shared toggle condition into input/helpers.rs so both popups (and
any future one) use the same logic instead of copy-pasting it.
