+++ src/tui/logic/input/quit.rs
@@ -1,16 +1,13 @@
 use ratatui::crossterm::event::{KeyCode, KeyEvent};
 
 use super::InputOutcome;
+use super::helpers::toggles_binary_choice;
 use crate::tui::logic::state::AppState;
 
 pub(crate) fn handle_quit_confirm(key: KeyEvent, state: &mut AppState) -> InputOutcome {
     match key.code {
         KeyCode::Esc => {
             state.quit_confirm_visible = false;
-            InputOutcome::Continue
-        }
-        KeyCode::Left | KeyCode::Right | KeyCode::Up | KeyCode::Down => {
-            state.quit_confirm_selected = if state.quit_confirm_selected == 0 { 1 } else { 0 };
             InputOutcome::Continue
         }
         KeyCode::Enter => {
@@ -20,6 +17,10 @@ pub(crate) fn handle_quit_confirm(key: KeyEvent, state: &mut AppState) -> InputO
                 state.quit_confirm_visible = false;
                 InputOutcome::Continue
             }
+        }
+        _ if toggles_binary_choice(&key, &state.keymap) => {
+            state.quit_confirm_selected = if state.quit_confirm_selected == 0 { 1 } else { 0 };
+            InputOutcome::Continue
         }
         _ => InputOutcome::Continue,
     }
